feat: implement fix enum data request and response

+++ FoodHub.Tests/Features/MenuItems/CreateMenuItemHandlerTests.cs
@@ -48,7 +48,7 @@ namespace FoodHub.Tests.Features.MenuItems
                 "https://example.com/image.jpg",
                 "Phở bò truyền thống",
                 categoryId,
-                Station.HotKitchen,
+                (int)Station.HotKitchen,
                 15,
                 50000m,
                 25000m
@@ -106,7 +106,7 @@ namespace FoodHub.Tests.Features.MenuItems
                 null,
                 null,
                 categoryId,
-                Station.HotKitchen,
+                (int)Station.HotKitchen,
                 15,
                 50000m,
                 null
@@ -155,7 +155,7 @@ namespace FoodHub.Tests.Features.MenuItems
                 null,
                 null,
                 categoryId,
-                Station.HotKitchen,
+                (int)Station.HotKitchen,
                 15,
                 50000m,
                 null
